Fixed compilier warnings. For 4.0 beta release.

+++ hdf/src/dfufp2i.c
@@ -119,7 +119,7 @@ nduif2i(int32 *hdim, int32 *vdim, float32 *max, float32 *min, float32 hscale[],
     int         ret;
 
     fn = HDf2cstring(outfile, *lenfn);
-    ret = DFUfptoimage(*hdim, *vdim, *max, *min, hscale, vscale, data,
+    ret = DFUfptoimage(*hdim, *vdim, (float32)*max, (float32)*min, hscale, vscale, data,
                 _fcdtocp(palette), fn, *ct_method, *hres, *vres, *compress);
     HDfree(fn);
     return (ret);
